# specs/001-quickflow-productivity/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QuickFlow API
  version: 1.0.0
  description: >
    REST API of QuickFlow (single-user personal productivity). Source of truth for the backend springdoc swagger
    and the generated typescript-angular client. Errors use RFC 9457 application/problem+json.
servers:
- {url: 'http://localhost:8080'}
...
paths:
...
  /api/plans:
    get:
      tags: [Plans]
      operationId: listPlans
      summary: All plans with derived status, progress and rest time
      parameters:
      - name: status
        in: query
        required: false
        schema: {$ref: '#/components/schemas/PlanStatus'}
      responses:
        '200':
          description: Plans ordered by priorityOrder asc, then startDateTime asc
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Plan'}
        '400': {$ref: '#/components/responses/BadRequest'}
    post:
      tags: [Plans]
      operationId: createPlan
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PlanRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Plan'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/plans/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    get:
      tags: [Plans]
      operationId: getPlan
      responses:
        '200':
          description: The plan
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Plan'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Plans]
      operationId: deletePlan
      responses:
        '204': {description: Deleted with its items; sources untouched}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/plans/{id}/items/{itemId}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    - name: itemId
      in: path
      required: true
      schema: {type: integer, format: int64}
    put:
      tags: [Plans]
      operationId: setPlanItemDone
      summary: Mark a plan item done / not done (task/habit side effects per business rule 13)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PlanItemUpdateRequest'}
      responses:
        '200':
          description: Updated; returns the plan
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Plan'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/plans/{id}/start-notification:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Plans]
      operationId: acknowledgePlanStart
      summary: Record that the start notification was shown (idempotent)
      responses:
        '200':
          description: Acknowledged; returns the plan
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Plan'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/settings:
    get:
      tags: [Settings]
      operationId: getSettings
      responses:
        '200':
          description: Current settings (defaults on first call)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Settings'}
    put:
      tags: [Settings]
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Settings'}
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Settings'}
        '400': {$ref: '#/components/responses/BadRequest'}
components:
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        errors:
          type: array
          items: {$ref: '#/components/schemas/FieldError'}
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: {type: string}
        message: {type: string}
...
    PlanStatus:
      type: string
      enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
    PlanItemSourceType:
      type: string
      enum: [TASK, HABIT, LEARNING_RESOURCE]
    PlanItemSource:
      type: object
      required: [sourceType, sourceId]
      properties:
        sourceType: {$ref: '#/components/schemas/PlanItemSourceType'}
        sourceId: {type: integer, format: int64}
    PlanRequest:
      type: object
      required: [title, estimatedDurationMinutes, startDateTime, endDateTime, priorityOrder, items]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        estimatedDurationMinutes: {type: integer, minimum: 1}
        startDateTime: {type: string, format: date-time}
        endDateTime: {type: string, format: date-time, description: Must be after startDateTime}
        priorityOrder: {type: integer, minimum: 1}
        items:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/PlanItemSource'}
    PlanItemUpdateRequest:
      type: object
      required: [done]
      properties:
        done: {type: boolean}
    PlanItem:
      type: object
      required: [id, sourceType, sourceId, sourceTitle, done, sourceAvailable]
      properties:
        id: {type: integer, format: int64}
        sourceType: {$ref: '#/components/schemas/PlanItemSourceType'}
        sourceId: {type: integer, format: int64}
        sourceTitle: {type: string}
        done: {type: boolean}
        sourceAvailable: {type: boolean}
    Plan:
      type: object
      required: [id, title, estimatedDurationMinutes, startDateTime, endDateTime, priorityOrder, status, createdAt,
        items, itemsTotal, itemsDone, progressPercent]
      properties:
        id: {type: integer, format: int64}
        title: {type: string}
        estimatedDurationMinutes: {type: integer}
        startDateTime: {type: string, format: date-time}
        endDateTime: {type: string, format: date-time}
        priorityOrder: {type: integer}
        status: {$ref: '#/components/schemas/PlanStatus'}
        createdAt: {type: string, format: date-time}
        startNotifiedAt:
          type: [string, 'null']
          format: date-time
        items:
          type: array
          items: {$ref: '#/components/schemas/PlanItem'}
        itemsTotal: {type: integer}
        itemsDone: {type: integer}
        progressPercent: {type: integer, minimum: 0, maximum: 100}
        restSeconds:
          type: [integer, 'null']
          format: int64
          description: Seconds until endDateTime; only when IN_PROGRESS
...
    DefaultView:
      type: string
      enum: [DASHBOARD, TASKS, HABITS, LEARNING, PLANS]
    Settings:
      type: object
      required: [displayName, inAppNotifications, browserNotifications, defaultView]
      properties:
        displayName: {type: string, minLength: 1, maxLength: 80}
        inAppNotifications: {type: boolean}
        browserNotifications: {type: boolean}
        defaultView: {$ref: '#/components/schemas/DefaultView'}